# [Improve] pool sendRequest to improve producer perf