@Gleiphir2769 2023-11-06
Hi @gunli, the implement looks good to me. But most member of `SendRequest` are pointer. I think It's not essential to pool it. Could you give a more significant benchmark result?