@gunli 2023-11-06
> Hi @gunli, the implement looks good to me. But most member of `SendRequest` are pointer. I think It's not essential to pool it. Could you give a more significant benchmark result?

![heap-pr](https://github.com/apache/pulsar-client-go/assets/10734543/5244d3b6-61a7-4c25-a1f9-1a38e5c2c977)
![heap-nopr](https://github.com/apache/pulsar-client-go/assets/10734543/14a406de-9c82-485e-9593-21064a054454)

Yes, here are the heap pprof results when benchmarked with and without this PR, we can see when benchmarked  without this PR, `internalSend` and `internalSendAsync`, in where `sendRequest`s are allocated, allocated much more memory than that when it benchmarked with this PR.