@Gleiphir2769 2023-11-06
>  we can see when benchmarked without this PR, internalSend and internalSendAsync, in where sendRequests are allocated, allocated much more memory than that when it benchmarked with this PR.
Good work @gunli ! I have no doubt now.