@gunli 2023-11-08
For the sake of fairness, I updated the benchmark, set b.N = 500000, the pprof reports are:

![pr](https://github.com/apache/pulsar-client-go/assets/10734543/7fcb58fd-3627-41f7-b296-c3b1c70c14fe)
![no-pr](https://github.com/apache/pulsar-client-go/assets/10734543/a08bc650-091c-4267-aae3-80aed04870f8)